requirements.txt
aletk>=0.1.6
aiohttp>=3.11.13
pillow>=11.1.0
piexif>=1.1.3
google-cloud-vision>=3.10.1
google-auth>=2.38.0
playwright>=1.51.0
imagehash>=4.3.2
ipykernel>=6.29.5
black>=25.1.0
mypy>=1.15.0
mypy-extensions>=1.0.0